move query to own file
